+++ src/types/settings.d.ts
@@ -9,7 +9,7 @@ export type ToastPosition =
     | 'bottom-center'
     | 'bottom-right';
 
-export type SettingsContextProperties = {
+export type SettingsData = {
     ui: {
         toast: {
             duration: number;
@@ -24,5 +24,22 @@ export type SettingsContextProperties = {
             editEdge: boolean;
             import: boolean;
         };
+        limits: {
+            maxNodes: number;
+            maxEdges: number;
+        };
+    };
+};
+
+export type SettingsContextProperties = {
+    ui: {
+        toast: SettingsData['ui']['toast'];
+        setToast: (toast: SettingsData['ui']['toast']) => void;
+    };
+    graph: {
+        arrangeOn: SettingsData['graph']['arrangeOn'];
+        setArrangeOn: (arrangeOn: SettingsData['graph']['arrangeOn']) => void;
+        limits: SettingsData['graph']['limits'];
+        setLimits: (limits: SettingsData['graph']['limits']) => void;
     };
 };